+++ packages/MagicBase/src/base/basis.jl
@@ -359,3 +359,10 @@ function metroplis1d(loglfun::Function;
     chain
 end
 
+function exportall(mod)
+    for n in names(mod, all = true)
+        if Base.isidentifier(n) && n ∉ (Symbol(mod), :eval)
+            @eval mod export $n
+        end
+    end
+end